optimize seq2seq-dataset
